feat: add registerTypeBoxFormats helper to index
